Them xoa cho quan ly danh muc, quan ly vaccine

+++ app/Http/Controllers/DistrictController.php
@@ -37,6 +37,23 @@ class DistrictController extends DistrictRepository
         }
     }
 
+    public function getDataWithTrashed () {
+        $current_user = User::find(Auth::user()->id);
+
+        if ($current_user->can('viewAny', User::class)) {
+            $districts = $this->getDistrictsWithTrashed();
+
+            return view('admin.district.trash', [
+                'districts' => $districts,
+                'sort' => empty(request()->query('sort')) ? '' : request()->query('sort'),
+                'status' => empty(request()->query('status')) ? '' : request()->query('status'),
+                'name' => empty(request()->query('name')) ? '' : request()->query('name'),
+            ]);
+        } else {
+            return redirect()->route('admin.errors.4xx');
+        }
+    }
+
     /**
      * Show the form for creating a new resource.
      *
@@ -149,43 +166,49 @@ class DistrictController extends DistrictRepository
      */
     public function destroy($id)
     {
-        if ($this->deleteModel($id)) {
-            return response()->json(['mess' => 'Xóa bản ghi thành công'], 200);
+        $currentUser = User::find(Auth::user()->id);
+
+        if ($currentUser->can('viewAny', User::class)) {
+            if ($this->deleteModel($id)) {
+                return response()->json(['mess' => 'Xóa bản ghi thành công'], 200);
+            } else {
+                return response()->json(['mess' => 'Xóa bản không thành công'], 400);
+            }
         } else {
-            return response()->json(['mess' => 'Xóa bản không thành công'], 400);
+            return response()->json(['mess' => 'Xóa bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
         }
     }
 
     public function forceDelete($id)
     {
-        // $currentUser = User::findOrFail(Auth()->user()->id);
+        $currentUser = User::findOrFail(Auth()->user()->id);
 
-        // if ($currentUser->can('forceDelete', ClassRoom::class)) {
+        if ($currentUser->can('viewAny', User::class)) {
 
             if ($this->forceDeleteModel($id)) {
                 return response()->json(['mess' => 'Xóa bản ghi thành công'], 200);
             } else {
                 return response()->json(['mess' => 'Xóa bản không thành công'], 400);
             }
-        // } else {
-        //     return response()->json(['mess' => 'Xóa bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
-        // }
+        } else {
+            return response()->json(['mess' => 'Xóa bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
+        }
     }
 
     public function restore($id)
     {
-        // $currentUser = User::findOrFail(Auth()->user()->id);
+        $currentUser = User::findOrFail(Auth()->user()->id);
 
-        // if ($currentUser->can('restore', ClassRoom::class)) {
+        if ($currentUser->can('viewAny', User::class)) {
 
             if ($this->restoreModel($id)) {
-                return response()->json(['mess' => 'Khôi bản ghi thành công'], 200);
+                return response()->json(['mess' => 'Khôi phục bản ghi thành công'], 200);
             } else {
-                return response()->json(['mess' => 'Khôi bản không thành công'], 400);
+                return response()->json(['mess' => 'Khôi phục bản không thành công'], 400);
             }
-        // } else {
-        //     return response()->json(['mess' => 'Khôi phục bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
-        // }
+        } else {
+            return response()->json(['mess' => 'Khôi phục bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
+        }
     }
 
     public function getDistrictsByProvinceId ($id) {
