+++ app/Http/Controllers/ProvinceController.php
@@ -36,6 +36,23 @@ class ProvinceController extends ProvinceRepository
         }
     }
 
+    public function getDataWithTrashed () {
+        $current_user = User::find(Auth::user()->id);
+
+        if ($current_user->can('viewAny', User::class)) {
+            $provinces = $this->getProvincesWithTrashed();
+
+            return view('admin.province.trash', [
+                'provinces' => $provinces,
+                'sort' => empty(request()->query('sort')) ? '' : request()->query('sort'),
+                'status' => empty(request()->query('status')) ? '' : request()->query('status'),
+                'name' => empty(request()->query('name')) ? '' : request()->query('name'),
+            ]);
+        } else {
+            return redirect()->route('admin.errors.4xx');
+        }
+    }
+
     /**
      * Show the form for creating a new resource.
      *
@@ -108,8 +125,6 @@ class ProvinceController extends ProvinceRepository
         } else {
             return redirect()->route('admin.errors.4xx');
         }
-
-        
     }
 
     /**
@@ -132,7 +147,6 @@ class ProvinceController extends ProvinceRepository
         } else {
             return response()->json(['mess' => 'Sửa bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
         }
-        
     }
 
     /**
@@ -143,50 +157,48 @@ class ProvinceController extends ProvinceRepository
      */
     public function destroy($id)
     {
-        // $current_user = User::find(Auth::user()->id);
+        $currentUser = User::find(Auth::user()->id);
 
-        // if ($current_user->can('viewAny', User::class)) {
-        //     return view('admin.province.create');
-        // } else {
-        //     return redirect()->route('admin.errors.4xx');
-        // }
-
-        if ($this->deleteModel($id)) {
-            return response()->json(['mess' => 'Xóa bản ghi thành công'], 200);
+        if ($currentUser->can('viewAny', User::class)) {
+            if ($this->deleteModel($id)) {
+                return response()->json(['mess' => 'Xóa bản ghi thành công'], 200);
+            } else {
+                return response()->json(['mess' => 'Xóa bản không thành công'], 400);
+            }
         } else {
-            return response()->json(['mess' => 'Xóa bản không thành công'], 400);
+            return response()->json(['mess' => 'Xóa bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
         }
     }
 
     public function forceDelete($id)
     {
-        // $currentUser = User::findOrFail(Auth()->user()->id);
+        $currentUser = User::findOrFail(Auth()->user()->id);
 
-        // if ($currentUser->can('forceDelete', ClassRoom::class)) {
+        if ($currentUser->can('viewAny', User::class)) {
 
             if ($this->forceDeleteModel($id)) {
                 return response()->json(['mess' => 'Xóa bản ghi thành công'], 200);
             } else {
                 return response()->json(['mess' => 'Xóa bản không thành công'], 400);
             }
-        // } else {
-        //     return response()->json(['mess' => 'Xóa bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
-        // }
+        } else {
+            return response()->json(['mess' => 'Xóa bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
+        }
     }
 
     public function restore($id)
     {
-        // $currentUser = User::findOrFail(Auth()->user()->id);
+        $currentUser = User::findOrFail(Auth()->user()->id);
 
-        // if ($currentUser->can('restore', ClassRoom::class)) {
+        if ($currentUser->can('viewAny', User::class)) {
 
             if ($this->restoreModel($id)) {
-                return response()->json(['mess' => 'Khôi bản ghi thành công'], 200);
+                return response()->json(['mess' => 'Khôi phục bản ghi thành công'], 200);
             } else {
-                return response()->json(['mess' => 'Khôi bản không thành công'], 400);
+                return response()->json(['mess' => 'Khôi phục bản không thành công'], 400);
             }
-        // } else {
-        //     return response()->json(['mess' => 'Khôi phục bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
-        // }
+        } else {
+            return response()->json(['mess' => 'Khôi phục bản ghi lỗi, bạn không đủ thẩm quyền'], 403);
+        }
     }
 }
